move default pipelines of cores sources into source folders
