+++ dlt/sources/filesystem/default_pipeline.py
