+++ tests/sources/filesystem/test_filesystem_pipeline_template.py
@@ -15,8 +15,8 @@ from tests.common.storages.utils import TEST_SAMPLE_FILES
     ),
 )
 def test_all_examples(example_name: str) -> None:
-    from dlt.sources import filesystem_pipeline
+    from dlt.sources.filesystem import default_pipeline
 
-    filesystem_pipeline.TESTS_BUCKET_URL = TEST_SAMPLE_FILES
+    default_pipeline.TESTS_BUCKET_URL = TEST_SAMPLE_FILES
 
-    getattr(filesystem_pipeline, example_name)()
+    getattr(default_pipeline, example_name)()
