+++ tests/sources/rest_api/test_rest_api_pipeline_template.py
@@ -12,7 +12,7 @@ from dlt.common.typing import TSecretStrValue
     ),
 )
 def test_all_examples(example_name: str) -> None:
-    from dlt.sources import rest_api_pipeline
+    from dlt.sources.rest_api import default_pipeline
 
     # reroute token location from secrets
     github_token: TSecretStrValue = dlt.secrets.get("sources.github.access_token")
@@ -20,4 +20,4 @@ def test_all_examples(example_name: str) -> None:
         # try to get GITHUB TOKEN which is available on github actions, fallback to None if not available
         github_token = os.environ.get("GITHUB_TOKEN", None)  # type: ignore
     dlt.secrets["sources.rest_api_pipeline.github.access_token"] = github_token
-    getattr(rest_api_pipeline, example_name)()
+    getattr(default_pipeline, example_name)()
